requirements.txt
Pypdf2
streamlit
openai
python-dotenv